Appended PATCH call with Update analysis Connected Analysis data with Entry Update Data in prisma by Editor Component
