+++ journal/app/(dashboard)/journal/page.tsx
@@ -10,6 +10,9 @@ const getEntries = async () => {
     where: {
         userId: user.id,
     },
+    include: {
+        analysis: true
+    },
     orderBy: {
       createdAt: "desc",
     },
